+++ CHANGES.rst
@@ -51,6 +51,7 @@ Changed
 -------
 
 * Improve LINEST() compatibilty w/ Excel
+* Improve TEXT() compatibilty w/ Excel
 * Improve error and number handling in some Text functions
 * Improve IFS() to support array context
 * Missing references from INDIRECT() and OFFSET() resolve more often
@@ -61,7 +62,8 @@ Fixed
 * Fix #111, Incorrect implementation of YEARFRAC
 * Fixed some exceptions in LINEST()
 * Fix serialize ranges with formulas
-
+* Fixed a minor bug in DATE()
+* Fixed TIMEVALUE() parsing for elapsed times
 
 [1.0b26] - 2021-06-18
 =====================